rename(domain): CustomFirestore 클래스명을 추상클래스 변경하며 AbstractFirestore로 변경
