+++ Domains/hooks/Firestore/AbstractFirestore.ts
@@ -1,0 +1,6 @@
+abstract class AbstractFirestore {
+  constructor() {}
+  fetchData() {}
+}
+
+export default AbstractFirestore;
