--- Domains/hooks/Firestore/CustomFirestore.ts
@@ -1,5 +1,0 @@
-class CustomFirestore {
-  constructor() {}
-}
-
-export default CustomFirestore;
